Revert bf16 autocast (WSL2 CUDA bug is not precision-related)
bf16 still crashes — the issue is async kernel execution on WSL2,
not fp16 overflow. Revert to default autocast dtype. Fix is to
disable AMP entirely via config (enable_amp = false).

+++ mortal/train_awr.py
@@ -265,7 +265,7 @@ def train():
                     return
 
             try:
-                with torch.autocast(device.type, enabled=enable_amp, dtype=torch.bfloat16):
+                with torch.autocast(device.type, enabled=enable_amp):
                     phi = mortal(obs, invisible_obs)
                     probs = policy_net(phi, masks)
                     dist = Categorical(probs=probs)
